fix(entities): Fix to change Double types to BigDecimal.

+++ src/main/java/canabarro/matheus/canispendapi/entity/PreferencesEntity.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -22,7 +23,7 @@ public class PreferencesEntity {
     private Long id;
 
     @Column(name = "SAVING_PERCENTAGE", nullable = false)
-    private Double savingPercentage;
+    private BigDecimal savingPercentage;
 
     @Column(name = "UPDATED_AT", nullable = false)
     private LocalDateTime updatedAt;
