+++ src/main/java/canabarro/matheus/canispendapi/entity/SalaryEntity.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -22,7 +23,7 @@ public class SalaryEntity {
     private Long id;
 
     @Column(name = "INCOME", nullable = false)
-    private Double income;
+    private BigDecimal income;
 
     @Column(name = "PAY_DAY")
     private Byte payDay;
